+++ app/quotes/generator.py
@@ -1,11 +1,13 @@
+from sqlalchemy.orm import Session
+
 from app.quotes.calculator import QuoteCalculator
 
 class QuoteGenerator:
 
-    def generate(self, request):
+    def generate(self, db: Session, request):
 
         calculator = QuoteCalculator()
 
-        quote = calculator.calculate(request)
+        quote = calculator.calculate(db, request)
 
         return quote